+++ src/Misc/Hooks.LaserDraw.cpp
@@ -1,3 +1,4 @@
+#include <Ext\WeaponType\Body.h>
 #include <Helpers/Macro.h>
 #include <Utilities/GeneralUtils.h>
 
@@ -40,4 +41,15 @@ DEFINE_HOOK(0x550F47, LaserDrawClass_DrawInHouseColor_BetterDrawing, 0x0)
 	R->EDX(b);
 
 	return 0x550F9D;
+}
+
+DEFINE_HOOK(0x6FD3FD, TechnoClass_LaserZap_ZAdjust, 0x5)
+{
+	GET_STACK(WeaponTypeClass*, pWeapon, STACK_OFFSET(0x6C, 0xC));
+	GET(int, zAdjust, EAX);
+
+	zAdjust += WeaponTypeExt::ExtMap.Find(pWeapon)->LaserZAdjust.Get(RulesExt::Global()->LaserZAdjust);
+	R->EAX(zAdjust);
+
+	return 0;
 }